fix another deadlock when shutting down
deadlock was caused when a chunk was being loaded in AsyncBlockAccess by a heightmap worker and a voxel worker at once, with the heightmap worker doing the work and the voxel worker blocked. since the voxel worker was shutting down first, it blocked forever since the server thread tasks for heightmap weren't cancelled yet.