be able to cancel top-level recursive tasks efficiently
basically, if a recursive task is cancelled while executing, it's now able to recurse all the dependencies and release them as well, freeing up the worker almost immediately instead of leaving it doing useless work until the whole task is complete. and since tasks are reference-counted, it will still behave correctly in the case where one of the dependents is still required externally!