bump rocksdb dependency to 7.3.1
this is to work around the fact that version 6.28.2 of ForgeRocks on CurseForge doesn't have a minecraft version in its version string, thus screwing up the dependency check. bumping prevents people from downloading the wrong version, and i've removed the '-1.12.2' suffix requirement from the dependency declaration to prevent this from happening again.

this commit fixes #189.