core: initial api support for a 'rough FBlockLevel'
this will eventually be used to implement a standard rough generator, and merged rough+exact generator.

also, merged the vanilla and CubicChunks implementations of the regular exact generators for both heightmap and voxel mode. AbstractExactHeightmapGenerator and AbstractExactVoxelGenerator are now DefaultExact*Generator! having separate implementations was useful like a year ago, but by now i've added so much extra abstraction that the two implementations were completely identical. i've left the event around though, so that future versions or mods can override the standard exact generator if for some reason they want to do that.