mc:1.12.2-forge: limited support for using modded GenLayers with the IFastLayer pipeline
this has a lot of TODOs, but as it stands now this is able to wrap modded GenLayers of the "PaddedLayer" style and use them in the IFastLayer pipeline. the primary drawback is that now IFastLayer users have to manually check if the vanilla IntCache needs to be reset, but since that infrastructure is only ever referenced from a single class it seems an acceptable compromise.