mc:1.12.2-forge: make CWG noisegen roughly 12% faster
this takes advantage of the specific gradient vectors used by CWG to make the gradient vector array more compact, which reduces L1 cache thrashing while simultaneously allowing more efficient lookups to the gradient vector table. and to top it off, the resulting code is much, much cleaner! it's not every day you get an optimization which is a win/win/win in terms of storage, runtime and code readability :)