mc:1.12.2-forge: avoid NPE if fp2 hasn't been constructed yet
this avoids crashing the game when rendering a screen before the mod construction phase (which can happen if a mod dependency isn't satisfied). it should have negligible performance impact, although i will admit it's a pretty gross hack and i really need to come up with a better system for tracking the reversed-z state.