core: implement basic AIMD flow control using reported bandwidth from the client
the client now informs us of the time since the last tile data packet was received when it sends ACKs, which allows us to infer the effective bandwidth and adjust our sending speed so that we don't hog the entire network while sending tile data.

also rearranged packets a bit so that we use a static create() method instead of a constructor, in order to disambiguate from the no-args constructor which is only used internally.

also added a sessionId field which can be used to determine if a tile Ack packet belongs to a previously closed session and ignore it