mc:1.16-forge: crash when client thread throws exception during startup
Unfortunately, join()-ing the CompletableFuture for a task submitted to the client thread isn't a realistic solution, as doing so seems to cause the client thread to deadlock during mod initialization in certain situations. We have to resort to some gross hackery in order to force the game to crash.