core: implement debug statistics for GPU-culled render index
The displayed values are probably always going to be a few frames out-of-date by the time they actually show up on-screen, but that seems an acceptable tradeoff to avoid stalling the GPU pipeline. If that ever truly becomes a problem, I can add an option to force statistics to be downloaded synchronously.