gl: don't crash in AsynchronousSmallBufferDownloader when render-ahead limit is exceeded
In practice, this can be forced to occur reliably, at least on KDE+NVIDIA with unlimited framerate+VSync: when the game window is minimized, flipBuffers seems to no longer cause the CPU to wait for VSync, while the process seems to be demoted to lower priority on the GPU, resulting in the CPU easily getting hundreds of frames ahead of the GPU.

Just for shits and giggles, I've added a thing to periodically log when this occurs. This will later be improved once I have a unified system for aggregating performance warnings and notifying the user.