core: change all 1d compute shader workgroup sizes to 64
This is enough to fill exactly one wave on AMD devices, and two warps on Nvidia. In either case this should give the compiler more freedom with regards to register allocation, and so on, as smaller workgroup sizes means that all workgroup resources are freed up immediately when the warp completes.

(This should eventually be made dynamic based on GPU type, but for now this is good enough. I really wish OpenGL had some way to say "please choose the best workgroup size for me" for shaders with no communication between threads, though...)